requirements.txt
aiohttp==3.9.5
websockets==12.0
python-telegram-bot==20.8
aiosqlite==0.20.0
python-dotenv==1.0.1
